Added authentication to Journal and User Routes

+++ journalartifact/src/main/java/net/journalapp/journalartifact/service/JournalEntryService.java
@@ -7,9 +7,11 @@ import net.journalapp.journalartifact.repository.UserRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Component
 public class JournalEntryService {
@@ -23,12 +25,19 @@ public class JournalEntryService {
     @Autowired
     private UserService userService;
 
+    @Transactional
     public void saveEntry(JournalEntry journalEntry, String username){
-        User user = userService.findByUserName(username);
-        journalEntry.setDate(LocalDateTime.now());
-        JournalEntry saved = journalEntryRepository.save(journalEntry);
-        user.getJournalEntries().add(saved);
-        userService.saveEntry(user);
+        try{
+            User user = userService.findByUserName(username);
+            journalEntry.setDate(LocalDateTime.now());
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournalEntries().add(saved);
+            userService.saveUser(user);
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while saving the entry" , e);
+        }
+
     }
 
     //method overloading
@@ -36,19 +45,27 @@ public class JournalEntryService {
         journalEntryRepository.save(journalEntry);
     }
 
-    public List<JournalEntry> getAll(){
-        return journalEntryRepository.findAll();
+    public Optional<JournalEntry> findById(ObjectId id) {
+        return journalEntryRepository.findById(id);
     }
 
-    public JournalEntry getJournalByID(ObjectId id){
-        return journalEntryRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Journal Entry not found with ID: " + id));
-    }
+    @Transactional
+    public boolean deleteJournalByID(ObjectId id, String username){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(username);
+            removed =  user.getJournalEntries().removeIf(x->x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+                return true;
+            }
 
-    public void deleteJournalByID(ObjectId id, String username){
-        User user = userService.findByUserName(username);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
-        userRepository.save(user);
-        journalEntryRepository.deleteById(id);
+        }
+        catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry" , e);
+        }
+        return false;
     }
 }
